_deprecated/load_test/requirements.txt
aiohttp
python-dotenv
streamlit>=1.28.0
plotly

